Add calculations for FSRS retrievability for INPUT_MODE_SQLITE.
Retrievability output doesn't match the Anki `Stats` window
exactly, but it might not because it uses the system time when
calculating some date differences (That is, unlike other tables,
the results will change throughout the day and not just at the
rollover hour).

Also, minor fixes for bin labels so that '[5%, 10%)' sorts after
'[0%, 5%)' and not after '[45%, 50%)'.